Added electron stream, beam spread
Electrons are now created at one end of the apparatus as elements in an
ArrayList...
They are then removed from memory when they reach the other end of the
apparatus.
Eventually this will happen when they collide with the screen as well.
Added a bit of randomness to electron spawn velocity (crudely simulating
random distribution of speeds at the filament).
Added randomness to arrival y-coordinate (crudely simulating vertical
beam spread).
Now need to tie the arrival velocity to the linac voltage, and the
arrival y-velocity to the arrival y-coordinate (simulating the small
fanning of the beam in this direction (usually < 2mm).